listing saved after fetching from api call

+++ app/src/main/java/lazycoder21/droid/crypto/data/local/entity/CryptoDetailLocal.kt
@@ -8,7 +8,7 @@ data class CryptoDetailLocal(
     @PrimaryKey val symbol: String = "",
     val volume: String? = null,
     val askPrice: String? = null,
-    val at: Int? = null,
+    val at: Long? = null,
     val lowPrice: String? = null,
     val highPrice: String? = null,
     val openPrice: String? = null,
