+++ app/src/main/java/lazycoder21/droid/crypto/data/remote/model/CryptoListingsRemote.kt
@@ -7,7 +7,7 @@ data class CryptoListingsRemote(
         val volume: String? = null,
         val symbol: String? = null,
         val askPrice: String? = null,
-        val at: Int? = null,
+        val at: Long? = null,
         val lowPrice: String? = null,
         val highPrice: String? = null,
         val openPrice: String? = null,
